+++ source/keep_only_video_and_audio_streams/changelog.md
@@ -1,3 +1,6 @@
 
+**<span style="color:#56adda">0.0.2</span>**
+- added option to extract & convert all subst to srt files in original directory
+
 **<span style="color:#56adda">0.0.1</span>**
 - Initial version